do not duplicate mixins on file and folder, test creation and last modification works
